add auth features and 404 page

+++ driver/views.py
@@ -7,8 +7,11 @@ from .forms import DriverForm
 # Create your views here.
 
 def index(request):
-    form=DriverForm()
-    return render(request,'driver/index.html',{'form':form})
+    if request.user.is_authenticated:
+        form=DriverForm()
+        return render(request,'driver/index.html',{'form':form})
+    else:
+        return redirect("http://localhost:8000/home/404")
 
 def driver(request):
     if request.POST:
@@ -19,22 +22,31 @@ def driver(request):
             form=DriverForm()
             return render(request,'driver/index.html',{'form':form,'success' : success_message})
     else:
-        form=DriverForm()
-        error_message='Something went wrong error'
-        return render(request,'driver/index.html',{ 'form' : form ,'error':error_message})
+        if request.user.is_authenticated:
+            form=DriverForm()
+            error_message='Something went wrong error'
+            return render(request,'driver/index.html',{ 'form' : form ,'error':error_message})
+        else:
+            return redirect("http://localhost:8000/home/404")
 
 def drivers(request):
     if request.POST:
         form=DriverForm(request.POST)
         return render(request,'driver/index.html',{'form':form})
     else:
-        drivers = Driver.objects.all()
-        return render(request,'driver/driverlist.html',{ 'drivers' : drivers ,'user':request.user})
+        if request.user.is_authenticated:
+            drivers = Driver.objects.all()
+            return render(request,'driver/driverlist.html',{ 'drivers' : drivers ,'user':request.user})
+        else:
+            return redirect("http://localhost:8000/home/404")
 
 def delete(request,id):
     if request.POST:
         return render(request,'driver/index.html',{'form':form})
     else:
-        drivers = Driver.objects.get(id=id)
-        drivers.delete()
-        return redirect('http://localhost:8000/driver/drivers')
+        if request.user.is_authenticated:
+            drivers = Driver.objects.get(id=id)
+            drivers.delete()
+            return redirect('http://localhost:8000/driver/drivers')
+        else:
+            return redirect("http://localhost:8000/home/404")
